Refactor ffb_device_input with LowPassFilter composition, optimize LPF with batched alpha
- Move update_axis() and reset() implementations to ffb_device_input.cpp
- Compose with LowPassFilter instances instead of hand-rolled EMA
- Fix deadband ordering: derive speed/accel before deadband zeroing
- Clamp speed and acceleration to max bounds
- Remove volatile from Metrics (single-task-owned)
- Change Metrics maxPosition/maxSpeed/maxAcceleration to float
- Add DEFAULT_MAX_ACCEL_SCALE, retune DEFAULT_MAX_SPEED_SCALE 0.0025 -> 2.5
- Remove *1e3f from speed/accel calculations (units are now per-second)
- Optimize LowPassFilter: split operator() into overloads (explicit/auto dt)
- Add filter_alpha(x, alpha) for batched alpha computation
- Precompute alpha per axis group in update_axis(), reducing divisions

+++ src/ffb_force_calculator.cpp
@@ -115,7 +115,7 @@ namespace SunFFB
         return -force;
     }
 
-    void FFBForceCalculator::condition_force_calculator(const EffectBlock& effectBlock, const float metrics[NUM_AXIS], const int32_t maxMetrics[NUM_AXIS], float forces[NUM_AXIS]) const
+    void FFBForceCalculator::condition_force_calculator(const EffectBlock& effectBlock, const float metrics[NUM_AXIS], const float maxMetrics[NUM_AXIS], float forces[NUM_AXIS]) const
     {
         const uint8_t axisEnable = effectBlock.effectData.axisEnable;
         const uint8_t conditionBlockFlags = effectBlock.conditionBlockFlags;
